More archive.org related stuff
Add a script that checks what the spn.sh script's managed to grab so far, and emits a list that can be fed back in to start it again.
That works much better than it's built-in resume feature in the event of failure, as when it discovers a link's already archived, it wasn't counted as a success, and would be tried again when you resumed.

Also disable the email results as they're after every capture attempt, not a batch report at the end.
Finally, switch to the Wayback Machine's built-in outlinks system as spn.sh's system will work transitively until it's captured the whole internet, and ignores the separate outlink time limit.